test(frontend): split agent orchestrator operations suite
